safer files overwriting in case of undesired reboot

+++ save_snapshot.sh
@@ -4,9 +4,9 @@
 set -eo pipefail
 
 # Do not run script twice in parallel if it takes too much time
-LOCK=/tmp/gpu-monitor.lock
+LOCK=$(dirname $0)/data/gpu-monitor.lock
 if [ -e "$LOCK" ]; then
-  echo "Lock file already present, skipping run..."
+  echo "Lock file $LOCK already present, skipping run..."
   exit 1
 fi
 touch $LOCK
@@ -26,13 +26,23 @@ pyenv activate gpuview
 JSONFILE=data/json/$(date +%Y%m%d%H%M%S).json
 gpustat -a --json > $JSONFILE
 if test -s $JSONFILE; then
-  if ls data | grep .csv.gz > /dev/null; then
-    gunzip -k data/*.csv.gz
-  fi
+  find data -name *.csv.gz | while read f; do
+    gunzip -k $f
+  done
   python parse_json.py $JSONFILE
-  gzip -S .tmpgz data/*.csv
-  ls data/*.tmpgz | while read f; do
+  gzip -k -S .tmpgz data/*.csv
+  find data -name *.csv | while read f; do
+    mv $f $f.old
+  done
+  find data -name *.csv.tmpgz | while read f; do
+    if ! test -s $f; then
+      echo "GZIPped file $f is empty, there's something wrong, stopping here..."
+      exit 1
+    fi
     outf=${f/\.tmpgz/.gz}
+    if test -s $outf; then
+      mv $outf $outf.old
+    fi
     mv $f $outf
   done
